pyproject.toml
[project]
name = "floorplan-diffusion"
version = "0.1.0"
description = "Text-to-floorplan generation using diffusion models with LoRA fine-tuning"
authors = [
    {name = "FloorPlan Diffusion Team", email = "team@example.com"}
]
readme = "README.md"
        requires-python = ">=3.9"
dependencies = [
    # PyTorch with CUDA support (will be installed via pip with specific index)
    # "torch>=2.0.0",  # Commented out - will install separately
    # "torchvision>=0.15.0",  # Commented out - will install separately
    "torch",
    "torchvision",
    "transformers>=4.30.0",
    "diffusers>=0.21.0",
    "accelerate>=0.20.0",
    "peft>=0.4.0",
    "datasets>=2.12.0",
    "pillow>=9.5.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "scikit-learn>=1.3.0",
    "albumentations>=1.3.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

        [tool.pytest.ini_options]
        testpaths = ["tests"]
        python_files = ["test_*.py"]
        python_classes = ["Test*"]
        python_functions = ["test_*"]

        [tool.hatch.build.targets.wheel]
        packages = ["src"] 

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu126" }]
torchvision = [{ index = "pytorch-cu126" }]